refactored file structure
packetlib, server, and client now all live in seperate directores
